Move debug operator to separate file
