+++ src/debug-operator.ts
@@ -1,0 +1,23 @@
+import { Observable } from 'rxjs/Observable';
+import { tap } from 'rxjs/operators';
+import { environment } from './environments/environment';
+
+const debuggerOn = !environment.production;
+
+export const debugOperator = function<T> (message: string) {
+  const source = this;
+  if(debuggerOn) {
+    return source.pipe(tap(
+      function (next) {
+        console.log(message, next);
+      },
+      function (err) {
+        console.error('ERROR >>> ', message, err);
+      },
+      function () {
+        console.log(message, 'Completed.');
+      }
+    ));
+  }
+  return source;
+};
